requirements.txt
PySimpleGUI~=4.57.0
openpyxl~=3.0.9
Pillow~=9.0.1